Pull request main to jpv49

+++ CoreCare/Models/StressWorker.cs
@@ -11,38 +11,86 @@ namespace CoreCare.Models
     {
         private CancellationTokenSource? _cts;
 
+        private CancellationToken GetOrCreateToken()
+        {
+            if (_cts == null || _cts.IsCancellationRequested)
+            {
+                _cts = new CancellationTokenSource();
+            }
+
+            return _cts.Token;
+        }
+
         public void RunCpuStress(int seconds) //controla la duracion del estres, para no tenemos que llamar el stop manuelmente luego
         {
-            _cts = new CancellationTokenSource();
-            var token = _cts.Token;
+            var token = GetOrCreateToken();
             var end = DateTime.Now.AddSeconds(seconds);
+
             Task.Run(() =>
             {
+                var random = new Random();
                 while (DateTime.Now < end && !token.IsCancellationRequested)
-                    Math.Sqrt(new Random().NextDouble());
+                {
+                    _ = Math.Sqrt(random.NextDouble());
+                }
             }, token);
         }
 
         public void RunRamStress(int seconds)
         {
-            _cts = new CancellationTokenSource();
-            Task.Delay(seconds * 1000, _cts.Token).ContinueWith(_ => { });
-            var block = new byte[256 * 1024 * 1024];
-            Task.Delay(seconds * 1000).Wait();
-            GC.Collect();
+            var token = GetOrCreateToken();
+
+            Task.Run(async () =>
+            {
+                var end = DateTime.Now.AddSeconds(seconds);
+                var blocks = new List<byte[]>();
+
+                try
+                {
+                    while (DateTime.Now < end && !token.IsCancellationRequested)
+                    {
+                        blocks.Add(new byte[16 * 1024 * 1024]);
+
+                        if (blocks.Count >= 16)
+                        {
+                            await Task.Delay(250, token);
+                        }
+                    }
+                }
+                catch (OperationCanceledException)
+                {
+                }
+                finally
+                {
+                    blocks.Clear();
+                    GC.Collect();
+                }
+            }, token);
         }
 
         public void RunDiskStress(int seconds)
         {
-            _cts = new CancellationTokenSource();
-            var token = _cts.Token;
+            var token = GetOrCreateToken();
             var path = Path.GetTempFileName();
+
             Task.Run(() =>
             {
                 var end = DateTime.Now.AddSeconds(seconds);
-                while (DateTime.Now < end && !token.IsCancellationRequested)
-                    File.WriteAllBytes(path, new byte[1024 * 1024]);
-                File.Delete(path);
+
+                try
+                {
+                    while (DateTime.Now < end && !token.IsCancellationRequested)
+                    {
+                        File.WriteAllBytes(path, new byte[1024 * 1024]);
+                    }
+                }
+                finally
+                {
+                    if (File.Exists(path))
+                    {
+                        File.Delete(path);
+                    }
+                }
             }, token);
         }
